improve debug symbol discovery
If elfutils fails to find the debug symbols, the perfparser tries to
find it by checking first in /usr/lib/ than in .debug and than in
/usr/lib/debug. This causes some problems if the debug file has the same
name as the library itself. In this case the algorithm stops at
/usr/lib/ and returns the wrong file.
To prevent this I reorder the algorithm to first check in .debug than in
/usr/lib/debug and than in /usr/lib/

I also changed how the algorithm searches inside /usr/lib/debug. Some
distros (e.g. Ubuntu) uses shorted path names so the debug file for
/usr/lib/x86_64-linux-gnu/libc-2.31.so is located in
/usr/lib/debug/lib/x86_64-linux-gnu/libc-2.31.so. To find the file the
algorithm uses combines /usr/lib/debug/ with the path to the original
file. If the files does not exists it removes the first directory from
the path and tries again. (in case of the libc example the new path
would be /lib/x86_64-linux-gnu/libc-2.31.so)

fixes KDAB/hotspot#265
fixes KDAB/hotspot#128